+++ diesel/src/sqlite/backend.rs
@@ -61,7 +61,7 @@ impl SqlDialect for Sqlite {
     type ConcatClause = sql_dialect::concat_clause::ConcatWithPipesClause;
     type DefaultValueClauseForInsert = sql_dialect::default_value_clause::AnsiDefaultValueClause;
 
-    type BatchUpdateSupport = SqliteBatchUpdate;
+    type BatchUpdateSupport = sql_dialect::batch_update_support::DoesNotSupportBatchUpdate;
 
     type EmptyFromClauseSyntax = sql_dialect::from_clause_syntax::AnsiSqlFromClauseSyntax;
     type SelectStatementSyntax = sql_dialect::select_statement_syntax::AnsiSqlSelectStatement;
@@ -92,8 +92,6 @@ impl sql_dialect::on_conflict_clause::PgLikeOnConflictClause for SqliteOnConflic
 
 #[derive(Debug, Copy, Clone)]
 pub struct SqliteBatchInsert;
-#[derive(Debug, Copy, Clone)]
-pub struct SqliteBatchUpdate;
 
 #[derive(Debug, Copy, Clone)]
 pub struct SqliteReturningClause;